+++ src/basics/IfElse.sol
@@ -2,7 +2,7 @@
 pragma solidity 0.8.19;
 
 contract IfElse {
-    function ternary(uint _x) external pure returns(uint) {
+    function ternary(uint256 _x) external pure returns (uint256) {
         return _x >= 10 ? 1 : 11;
     }
 }
