+++ src/basics/Modif.sol
@@ -2,9 +2,9 @@
 pragma solidity 0.8.19;
 
 contract M {
-    uint count;
+    uint256 count;
 
-    modifier sandwich(uint c) {
+    modifier sandwich(uint256 c) {
         count += c;
         _;
         count *= 2;
@@ -13,5 +13,4 @@ contract M {
     function a() public sandwich(5) {
         count = 8;
     }
-
 }
