+++ src/basics/tests/GlobalVars.t.sol
@@ -5,7 +5,6 @@ import "lib/forge-std/src/Test.sol";
 import "../GlobalVars.sol";
 
 contract GlobalVarsTest is Test {
-
     GlobalVariables gvContract;
 
     function setUp() external {
@@ -13,11 +12,10 @@ contract GlobalVarsTest is Test {
     }
 
     function testGlobalVars() public view {
-        (address a, uint b, uint c) = gvContract.globalVars();
+        (address a, uint256 b, uint256 c) = gvContract.globalVars();
         console2.log("Global variables ");
         console2.log(a, b, c);
         assertEq(b, 1);
         assertEq(c, 1);
     }
-
 }
